Move create_version_for_upload to utils instead of tasks
