+++ src/olympia/devhub/tests/test_tasks.py
@@ -30,7 +30,6 @@ from olympia.devhub import tasks
 from olympia.files.models import File
 from olympia.files.utils import NoManifestFound
 from olympia.files.tests.test_models import UploadMixin
-from olympia.versions.models import Version
 
 
 pytestmark = pytest.mark.django_db
@@ -672,144 +671,6 @@ class TestSubmitFile(UploadMixin, TestCase):
         upload = self.get_upload(abspath=file_, addon=self.addon, version='1.0')
         tasks.submit_file(self.addon.pk, upload.pk, amo.RELEASE_CHANNEL_LISTED)
         assert not self.create_version_for_upload.called
-
-
-class TestCreateVersionForUpload(UploadMixin, TestCase):
-    fixtures = ['base/addon_3615']
-
-    def setUp(self):
-        super().setUp()
-        self.addon = Addon.objects.get(pk=3615)
-        self.mocks = {}
-        for key in ['Version.from_upload', 'parse_addon']:
-            patcher = mock.patch('olympia.devhub.tasks.%s' % key)
-            self.mocks[key] = patcher.start()
-            self.addCleanup(patcher.stop)
-        self.user = user_factory()
-
-    def test_file_passed_all_validations_not_most_recent(self):
-        file_ = get_addon_file('valid_webextension.xpi')
-        upload = self.get_upload(
-            abspath=file_, user=self.user, addon=self.addon, version='1.0'
-        )
-        newer_upload = self.get_upload(
-            abspath=file_, user=self.user, addon=self.addon, version='1.0'
-        )
-        newer_upload.update(created=datetime.today() + timedelta(hours=1))
-
-        # Check that the older file won't turn into a Version.
-        tasks.create_version_for_upload(self.addon, upload, amo.RELEASE_CHANNEL_LISTED)
-        assert not self.mocks['Version.from_upload'].called
-
-        # But the newer one will.
-        tasks.create_version_for_upload(
-            self.addon, newer_upload, amo.RELEASE_CHANNEL_LISTED
-        )
-        self.mocks['Version.from_upload'].assert_called_with(
-            newer_upload,
-            self.addon,
-            amo.RELEASE_CHANNEL_LISTED,
-            selected_apps=[amo.FIREFOX.id, amo.ANDROID.id],
-            parsed_data=self.mocks['parse_addon'].return_value,
-        )
-
-    def test_file_passed_all_validations_version_exists(self):
-        file_ = get_addon_file('valid_webextension.xpi')
-        upload = self.get_upload(
-            abspath=file_, user=self.user, addon=self.addon, version='1.0'
-        )
-        Version.objects.create(addon=upload.addon, version=upload.version)
-
-        # Check that the older file won't turn into a Version.
-        tasks.create_version_for_upload(self.addon, upload, amo.RELEASE_CHANNEL_LISTED)
-        assert not self.mocks['Version.from_upload'].called
-
-    def test_file_passed_all_validations_most_recent_failed(self):
-        file_ = get_addon_file('valid_webextension.xpi')
-        upload = self.get_upload(
-            abspath=file_, user=self.user, addon=self.addon, version='1.0'
-        )
-        newer_upload = self.get_upload(
-            abspath=file_, user=self.user, addon=self.addon, version='1.0'
-        )
-        newer_upload.update(
-            created=datetime.today() + timedelta(hours=1),
-            valid=False,
-            validation=json.dumps({'errors': 5}),
-        )
-
-        tasks.create_version_for_upload(self.addon, upload, amo.RELEASE_CHANNEL_LISTED)
-        assert not self.mocks['Version.from_upload'].called
-
-    def test_file_passed_all_validations_most_recent(self):
-        file_ = get_addon_file('valid_webextension.xpi')
-        upload = self.get_upload(
-            abspath=file_, user=self.user, addon=self.addon, version='1.0'
-        )
-        newer_upload = self.get_upload(
-            abspath=file_, user=self.user, addon=self.addon, version='0.5'
-        )
-        newer_upload.update(created=datetime.today() + timedelta(hours=1))
-
-        # The Version is created because the newer upload is for a different
-        # version_string.
-        tasks.create_version_for_upload(self.addon, upload, amo.RELEASE_CHANNEL_LISTED)
-        self.mocks['parse_addon'].assert_called_with(upload, self.addon, user=self.user)
-        self.mocks['Version.from_upload'].assert_called_with(
-            upload,
-            self.addon,
-            amo.RELEASE_CHANNEL_LISTED,
-            selected_apps=[amo.FIREFOX.id, amo.ANDROID.id],
-            parsed_data=self.mocks['parse_addon'].return_value,
-        )
-
-    def test_file_passed_all_validations_beta_string(self):
-        file_ = get_addon_file('valid_webextension.xpi')
-        upload = self.get_upload(
-            abspath=file_, user=self.user, addon=self.addon, version='1.0beta1'
-        )
-        tasks.create_version_for_upload(self.addon, upload, amo.RELEASE_CHANNEL_LISTED)
-        self.mocks['parse_addon'].assert_called_with(upload, self.addon, user=self.user)
-        self.mocks['Version.from_upload'].assert_called_with(
-            upload,
-            self.addon,
-            amo.RELEASE_CHANNEL_LISTED,
-            selected_apps=[amo.FIREFOX.id, amo.ANDROID.id],
-            parsed_data=self.mocks['parse_addon'].return_value,
-        )
-
-    def test_file_passed_all_validations_no_version(self):
-        file_ = get_addon_file('valid_webextension.xpi')
-        upload = self.get_upload(
-            abspath=file_, user=self.user, addon=self.addon, version=None
-        )
-        tasks.create_version_for_upload(self.addon, upload, amo.RELEASE_CHANNEL_LISTED)
-        self.mocks['parse_addon'].assert_called_with(upload, self.addon, user=self.user)
-        self.mocks['Version.from_upload'].assert_called_with(
-            upload,
-            self.addon,
-            amo.RELEASE_CHANNEL_LISTED,
-            selected_apps=[amo.FIREFOX.id, amo.ANDROID.id],
-            parsed_data=self.mocks['parse_addon'].return_value,
-        )
-
-    def test_pass_parsed_data(self):
-        file_ = get_addon_file('valid_webextension.xpi')
-        upload = self.get_upload(
-            abspath=file_, user=self.user, addon=self.addon, version=None
-        )
-        parsed_data = mock.Mock()
-        tasks.create_version_for_upload(
-            self.addon, upload, amo.RELEASE_CHANNEL_LISTED, parsed_data=parsed_data
-        )
-        assert self.mocks['parse_addon'].call_count == 0
-        self.mocks['Version.from_upload'].assert_called_with(
-            upload,
-            self.addon,
-            amo.RELEASE_CHANNEL_LISTED,
-            selected_apps=[amo.FIREFOX.id, amo.ANDROID.id],
-            parsed_data=parsed_data,
-        )
 
 
 class TestAPIKeyInSubmission(UploadMixin, TestCase):
