+++ src/olympia/devhub/utils.py
@@ -4,6 +4,7 @@ import waffle
 
 from celery import chain, chord
 from django.conf import settings
+from django.db import transaction
 from django.forms import ValidationError
 from django.utils.translation import gettext
 
@@ -16,6 +17,7 @@ from olympia.files.tasks import repack_fileupload
 from olympia.files.utils import parse_addon, parse_xpi
 from olympia.scanners.tasks import run_customs, run_wat, run_yara, call_mad_api
 from olympia.translations.models import Translation
+from olympia.versions.models import Version
 from olympia.versions.utils import process_color_value
 
 from . import tasks
@@ -351,3 +353,42 @@ def add_manifest_version_error(validation):
                 'fatal': True,
             },
         )
+
+
+@transaction.atomic
+def create_version_for_upload(addon, upload, channel, parsed_data=None):
+    fileupload_exists = addon.fileupload_set.filter(
+        created__gt=upload.created, version=upload.version
+    ).exists()
+    version_exists = Version.unfiltered.filter(
+        addon=addon, version=upload.version
+    ).exists()
+    if fileupload_exists or version_exists:
+        log.info(
+            'Skipping Version creation for {upload_uuid} that would '
+            ' cause duplicate version'.format(upload_uuid=upload.uuid)
+        )
+        return None
+    else:
+        log.info(
+            'Creating version for {upload_uuid} that passed '
+            'validation'.format(upload_uuid=upload.uuid)
+        )
+        # Note: if we somehow managed to get here with an invalid add-on,
+        # parse_addon() will raise ValidationError and the task will fail
+        # loudly in sentry.
+        if parsed_data is None:
+            parsed_data = parse_addon(upload, addon, user=upload.user)
+        version = Version.from_upload(
+            upload,
+            addon,
+            channel,
+            selected_apps=[x[0] for x in amo.APPS_CHOICES],
+            parsed_data=parsed_data,
+        )
+        # The add-on's status will be STATUS_NULL when its first version is
+        # created because the version has no files when it gets added and it
+        # gets flagged as invalid. We need to manually set the status.
+        if addon.status == amo.STATUS_NULL and channel == amo.RELEASE_CHANNEL_LISTED:
+            addon.update(status=amo.STATUS_NOMINATED)
+        return version
